chore: generalise vega market component
