--- frontend/components/receipts/utils/order/order-market.tsx
@@ -1,23 +1,0 @@
-import get from 'lodash/get'
-
-import { useMarketsStore } from '@/stores/markets-store'
-
-import { MarketLink } from '../../../vega-entities/market-link'
-
-export const locators = {
-  orderDetailsMarketCode: 'order-details-market-code'
-}
-
-export const OrderMarket = ({ marketId }: { marketId: string }) => {
-  const { getMarketById, loading } = useMarketsStore((state) => ({
-    loading: state.loading,
-    getMarketById: state.getMarketById
-  }))
-  const market = !loading && getMarketById(marketId)
-  const code = get(market, 'tradableInstrument.instrument.code')
-  if (!code) {
-    return <MarketLink key="order-details-market" marketId={marketId} />
-  }
-
-  return <div data-testid={locators.orderDetailsMarketCode}>{code}</div>
-}
